services/pq-crypto-service/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.0
pydantic==2.7.0
pydantic-settings==2.3.0
asyncpg==0.29.0
liboqs-python==0.10.1
cryptography==43.0.0
httpx==0.27.0
